1) Changes in CMakeLists.txt to handle Boost autolink on windows 2) Explicit casts to avoid compilation warnings with VC++ 3) Warnings are set to level 3 (/W3) to avoid STL and Boost compilation warnings with VC++ 4) In LidarIteratorEcho.h, line 302: friend class LidarConstIteratorEcho; --> friend struct LidarConstIteratorEcho; to avoid compilation warning with VC++

+++ src/LidarFormat/geometry/LidarCenteringTransfo.cpp
@@ -187,9 +187,9 @@ struct FunctorCenter
 		{
 			LidarEcho echoInitial(*itb);
 
-			echo.value<float>(decalageX) = echoInitial.value<AttributeType>(decalageX_initial) - transfo.x();
-			echo.value<float>(decalageY) = echoInitial.value<AttributeType>(decalageY_initial) - transfo.y();
-			echo.value<float>(decalageZ) = echoInitial.value<AttributeType>(decalageZ_initial);
+			echo.value<float>(decalageX) = (float)( echoInitial.value<AttributeType>(decalageX_initial) - transfo.x() );
+			echo.value<float>(decalageY) = (float)( echoInitial.value<AttributeType>(decalageY_initial) - transfo.y() );
+			echo.value<float>(decalageZ) = (float)echoInitial.value<AttributeType>(decalageZ_initial);
 
 			AttributeMapType::const_iterator itbAttribute = centeredContainer->getAttributeMap().begin();
 			const AttributeMapType::const_iterator iteAttribute = centeredContainer->getAttributeMap().end();
